feat: basic static gtfs caching with SQLite
Created connection handler with a SQLite database. Each GTFS entity has
its own table in the database, so that it's easier to retrieve data
without parsing the text files each time.

Since we are using ORMLite, each table is represented by a class model,
where each column of the table is a class field

Needs improvement by adding dedicated DatabaseService and GtfsService,
so that the code is more decoupled (but, unfortunately, more verbose)